The flash mode compares as one width on both iOS targets
AVCaptureFlashMode carries an isize on the simulator target, and NSNumber
hands back an i64, so the comparison did not build there.

+++ crates/cranpose/src/ios_camera.rs
@@ -385,7 +385,7 @@ fn capture_photo() -> Option<CameraStill> {
         };
         let supported = unsafe { holder.photo_output.supportedFlashModes() }
             .iter()
-            .any(|mode| mode.as_i64() == wanted.0);
+            .any(|mode| mode.as_i64() == wanted.0 as i64);
         if supported {
             unsafe { settings.setFlashMode(wanted) };
         }
